TCIAlgorithms

Documentation for TCIAlgorithms.

TCIAlgorithms.AbstractPatchCreator

TCIAlgorithms.BlockStructure

TCIAlgorithms.LazyMatrixMul

TCIAlgorithms.PatchOrdering

TCIAlgorithms.ProjTensorTrain

TCIAlgorithms.ProjTensorTrain

TCIAlgorithms.ProjTensorTrain

TCIAlgorithms.ProjectableEvaluator

TCIAlgorithms.ProjectableEvaluator

TCIAlgorithms.ProjectableEvaluatorAdapter

TCIAlgorithms._FuncAdapterTCI2Subset

TCIAlgorithms._performmul!

TCIAlgorithms.approxtt

TCIAlgorithms.batchevaluateprj

TCIAlgorithms.delete_node!

TCIAlgorithms.delete_value!

TCIAlgorithms.fulllength_leftmmultiidxset

TCIAlgorithms.fulltensor

TCIAlgorithms.fulltensor

TCIAlgorithms.maskactiveindices

TCIAlgorithms.matmul

TCIAlgorithms.project

TCIAlgorithms.project_on_subsetsiteinds

TCIAlgorithms.AbstractPatchCreator — Type

T: Float64, ComplexF64, etc. M: TensorCI2, MPS, etc.

TCIAlgorithms.BlockStructure — Type

Consists of a collection of non-overlapping blocks. The blocks do not nessarily have to cover the whole space.

TCIAlgorithms.LazyMatrixMul — Type

Lazy evaluation for matrix multiplication of two TTOs

This is for the lazy evaluation of the matrix multiplication of two TTOs. If underlying data needs to be changed by projection, it will be copied.

TODO: make immutable

TCIAlgorithms.PatchOrdering — Type

Specify the ordering of patching

TCIAlgorithms.ProjTensorTrain — Type

Represents a TensorTrain object which can be projected on a subset of indices

Compared to TCI.TensorTrain, this object has additional functionalities:

Projection

Multi site indices per tensor are supported.

Fast evaluation by caching intermediate objects of contraction

TCIAlgorithms.ProjTensorTrain — Method

The user must make sure that the data is compatible with the given projector. We do not recommend to use this function directly. Use project(ProjTensorTrain(tt), projector) instead.

TCIAlgorithms.ProjTensorTrain — Method

tt is a TensorTrain{T,3} and prj is a Projector. tt is defined on unprojected indices. Return a ProjTensorTrain{T} defined on full indices.

TCIAlgorithms.ProjectableEvaluator — Type

Represents an object that can be projected on a subset of indices

Attributes:

projector: Projector object

sitedims: Vector{Vector{Int}} of the dimensions of the local indices

TCIAlgorithms.ProjectableEvaluator — Method

This is similar to batchevaluateprj, but the evaluation is done on all M indices. In the returned array, the element evaluates to 0 for a indexset that is out of the projector.

TCIAlgorithms.ProjectableEvaluatorAdapter — Type

Convert function f to a ProjectableEvaluator object

TCIAlgorithms._FuncAdapterTCI2Subset — Type

For TCI2

f is a function that can be evaluated at full-length indices (including projected and non-projected indices). The wrapped function can be evaluated only on unprojected indices.

TCIAlgorithms._performmul! — Method

Perform matrix multiplication of two tensor trains and if the bond dimension is too large, project the result to a lower bond dimension. We repeat this process until the bond dimension is small enough or no way to project more.

TCIAlgorithms.approxtt — Method

Compute a tensor train approximation of the object. The computation should be quick because the result will be used as initial guesses for the optimization. Override this function

TCIAlgorithms.batchevaluateprj — Method

Please override this funciton

This is similar to batch evaluation. The difference is as follows. If some of M central indices are projected, the evaluation is done on the projected indices. The sizes of the correponding indices in the returned array are set to 1.

leftmmultiidxset and rightmmultiidxset are defined for unprojected and projected indices.

TCIAlgorithms.delete_node! — Method

Remove a value at a given path

TCIAlgorithms.delete_value! — Method

Remove a value at a given path

TCIAlgorithms.fulllength_leftmmultiidxset — Method

leftmmultiidxset: Vector of indices on unprojected indices Returns: Vector of indices on projected and unprojected indices

TCIAlgorithms.fulltensor — Method

Evaluate obj at all possible indexsets and return a full tensor if reducesitedims is true, in the returned tensor, the dimensions of the projected site indices are reduced to 1.

TCIAlgorithms.fulltensor — Method

Evaluate obj at all possible indexsets and return a full tensor

TCIAlgorithms.maskactiveindices — Method

n is the length of the prefix.

TCIAlgorithms.matmul — Method

Perform matrix multiplication of two tensor trains and project the result to a block structure.

TCIAlgorithms.project — Method

Project the object on the overlap of prj and obj.projector.

The requirement for the implementation is that the projector of the returned object is a subset of prj.

TCIAlgorithms.project_on_subsetsiteinds — Method

Project the tensor train on the subset of site indices. The returned tensor train object has the number of site indices equal to the number of the unprojected subset in the input. The dimension of the unprojected site indices are the same as the original ones, i.e., before the projection. On each tensor core in the input tensor train, site indices are either projected or not projected. The data will be copied, and in the returned tensor train, the site indices are fused.