TCIAlgorithms

Documentation for TCIAlgorithms.

TCIAlgorithms.AbstractPatchCreator

TCIAlgorithms.ElementwiseProduct

TCIAlgorithms.MatrixProduct

TCIAlgorithms.MatrixProductSum

TCIAlgorithms.PartitionedTensorTrain

TCIAlgorithms.PatchOrdering

TCIAlgorithms.ProjectableEvaluator

TCIAlgorithms.ProjectedTensorTrain

TCIAlgorithms.ProjectedTensorTrain

TCIAlgorithms.ProjectedTensorTrainProduct

TCIAlgorithms.deltaproduct

TCIAlgorithms.maskactiveindices

TCIAlgorithms.sum

TCIAlgorithms.AbstractPatchCreator — Type

T: Float64, ComplexF64, etc. M: TensorCI2, MPS, etc.

TCIAlgorithms.ElementwiseProduct — Type

Elementwise product of two tensor trains One site index on each site.

TCIAlgorithms.MatrixProduct — Type

Matrix product of two tensor trains Two site indices on each site.

TCIAlgorithms.MatrixProductSum — Type

Sum of matrix products of two tensor trains Two site indices on each site.

TCIAlgorithms.PartitionedTensorTrain — Type

Collection of ProjectableEvaluator objects

The underlying data will be copied when projected.

TCIAlgorithms.PatchOrdering — Type

Specify the ordering of patching

TCIAlgorithms.ProjectableEvaluator — Type

Type for an object that can be projected on a subset of indices

Attributes:

projector: Projector object

sitedims: Vector{Vector{Int}} of the dimensions of the local indices

TCIAlgorithms.ProjectedTensorTrain — Type

TensorTrain projected on a subset of indices

The underlying data will be copied when projected.

TCIAlgorithms.ProjectedTensorTrain — Method

tt is a TensorTrain{T,3} and prj is a Projector. tt is defined on unprojected indices. Return a ProjectedTensorTrain{T} defined on full indices.

TCIAlgorithms.ProjectedTensorTrainProduct — Type

Represents the product of two projected tensor trains

TCIAlgorithms.deltaproduct — Method

Elementwise product in one index: $C_{ijk} = A_{ij} B_{jk}$

TCIAlgorithms.maskactiveindices — Method

n is the length of the prefix.

TCIAlgorithms.sum — Method

Sum over external indices